+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/CHANGELOG.md
@@ -49,6 +49,11 @@ Notes](../../RELEASENOTES.md).
   limit now responds with HTTP 500.
   ([#7487](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7487))
 
+* Add `PrometheusHttpListenerOptions.ResourceConstantLabels` property to select
+  resource attributes to add to each metric as constant labels. Defaults to
+  `null` (no resource attributes are added as metric labels).
+  ([#7471](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7471))
+
 ## 1.16.0-beta.1
 
 Released 2026-Jun-10